+++ projects/ion/src/lib/sidebar/sidebar.module.ts
@@ -1,17 +1,12 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
-import { IonIconComponent } from '../icon/icon.component';
+import { IonIconModule } from '../icon/icon.module';
 import { SidebarGroupComponent } from './sidebar-group/sidebar-group.component';
 import { SidebarItemComponent } from './sidebar-item/sidebar-item.component';
 import { SidebarComponent } from './sidebar.component';
 
 @NgModule({
-  imports: [CommonModule],
-  declarations: [
-    SidebarComponent,
-    SidebarItemComponent,
-    SidebarGroupComponent,
-    IonIconComponent,
-  ],
+  imports: [CommonModule, IonIconModule],
+  declarations: [SidebarComponent, SidebarItemComponent, SidebarGroupComponent],
 })
 export class SidebarModule {}
